Adds Business and moves UnitOfWork.cs To Repository layer
