+++ OngProject/Startup.cs
@@ -14,9 +14,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using OngProject.Core.Business;
+using OngProject.Core.Interfaces;
 using OngProject.DataAccess;
-using OngProject.Services;
-using OngProject.Services.Interfaces;
+using OngProject.Repositories;
+using OngProject.Repositories.Interfaces;
 
 namespace OngProject
 {
@@ -32,7 +34,7 @@ namespace OngProject
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddScoped<IRoleService, RoleService>();
+            services.AddScoped<IRoleBusiness, RoleBusiness>();
             services.AddScoped<IUnitOfWork, UnitOfWork>();
             services.AddSession();
             services.AddControllers();
